Refactor RateFormModal into RateForm and RateModal.
