+++ js/src/components/shipping-rate-section/estimated-shipping-rates-card/add-rate-form-modal.js
@@ -1,12 +1,14 @@
 /**
  * External dependencies
  */
+import { Button } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import RateFormModal from './rate-form-modal.js';
+import RateForm from './rate-form.js';
+import RateModal from './rate-modal.js';
 
 /**
  * @typedef { import(".~/data/actions").CountryCode } CountryCode
@@ -22,15 +24,42 @@ import RateFormModal from './rate-form-modal.js';
  * @param {(values: ShippingRateGroup) => void} props.onSubmit Called with submitted value.
  * @param {() => void} props.onRequestClose Callback to close the modal.
  */
-const AddRateFormModal = ( props ) => {
+const AddRateFormModal = ( {
+	countryOptions,
+	initialValues,
+	onSubmit,
+	onRequestClose,
+} ) => {
 	return (
-		<RateFormModal
-			{ ...props }
-			submitButtonChildren={ __(
-				'Add shipping rate',
-				'google-listings-and-ads'
-			) }
-		/>
+		<RateForm
+			initialValues={ initialValues }
+			onSubmit={ onSubmit }
+			onRequestClose={ onRequestClose }
+		>
+			{ ( formProps ) => {
+				const { isValidForm, handleSubmit } = formProps;
+
+				return (
+					<RateModal
+						formProps={ formProps }
+						countryOptions={ countryOptions }
+						buttons={ [
+							<Button
+								key="submit"
+								isPrimary
+								disabled={ ! isValidForm }
+								onClick={ handleSubmit }
+							>
+								{ __(
+									'Add shipping rate',
+									'google-listings-and-ads'
+								) }
+							</Button>,
+						] }
+					/>
+				);
+			} }
+		</RateForm>
 	);
 };
